% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agg}
\alias{agg}
\title{Aggregate two parselists}
\usage{
agg(l1, l2)
}
\arguments{
\item{l1, l2}{Two parseLR lists}
}
\value{
A list
}
\description{
The function \code{agg} aggregates two parseLR lists.
}
\details{
It corresponds to the Miranda \verb{++} notation. The empty list \code{list()} is the
unit element for this function, meaning that if either of the two arguments
is an empty list then it will not be present in the result, and if both are
empty then the result will be the empty list.
}
\keyword{internal}
