% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{clear_store}
\alias{clear_store}
\title{Clear all stored values}
\usage{
clear_store()
}
\description{
Removes all variables stored using \code{store()}. Use this to clean up
between independent parser runs or in test teardown.
}
