% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concat}
\alias{concat}
\title{Concatenate two lists.}
\usage{
concat(l1, l2)
}
\arguments{
\item{l1, l2}{two \code{list} objects}
}
\value{
A \code{list}.
}
\description{
Make sure that empty lists disappear
}
\keyword{internal}
