% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{Disposable}
\alias{Disposable}
\title{Recognize and discard an element with disposable content}
\usage{
Disposable(is.comment = function(x) return(FALSE))
}
\arguments{
\item{is.comment}{A boolean function that recognizes a comment}
}
\description{
Disposable lines are those that are empty lines, as defined by \link{Spacer}
or lines that satisfy an is.comment() test, which is to be specified. By
default is.comment returns FALSE.
}
\examples{
starts_with_hash <- function(x) grepl("^#",x[1])
Disposable(is.comment = starts_with_hash) (c("# my comment", "Important text"))
Disposable() (c("            ", "    ", "Important text"))

}
