% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure.list}
\alias{ensure.list}
\title{Ensure that the structure of objects is preserved.}
\usage{
ensure.list(x)
}
\arguments{
\item{x}{Any R-object.}
}
\value{
A list.
}
\description{
After successful parsing the result must be present in a list to preserve
its structure. For example, a data frame or a matrix must be preserved.
}
\keyword{internal}
