% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{fail}
\alias{fail}
\title{The parser that always fails}
\usage{
fail()
}
\description{
While \code{succeed} parser never fails, \code{fail} always does, regardless
of the input vector. It, therefore, returns the empty list.
}
\section{Formal description}{


\verb{fail inp = [].}
}

\examples{
fail()(letters[1:5])

}
