% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{failed}
\alias{failed}
\title{Testing for parser failure}
\usage{
failed(o)
}
\arguments{
\item{o}{Output from a parser.}
}
\value{
A logical value.
}
\description{
Use this function to test whether your parser failed, for example in
unit testing of your parsers when writing a package.
}
\examples{
d <- (literal("A") \%then\% literal("B"))(c("A","A"))
d
failed(d)

}
\seealso{
\code{\link[=print.marker]{print.marker()}}
}
