% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fst}
\alias{fst}
\alias{snd}
\title{Selecting only left or right part from a \verb{\%then\%} sequence}
\usage{
fst(x)

snd(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector.
}
\description{
Used by \%xthen\% and \%thenx\%
}
\keyword{internal}
