% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{has_stored}
\alias{has_stored}
\title{Check if a variable is stored}
\usage{
has_stored(name)
}
\arguments{
\item{name}{Variable name to check}
}
\value{
Logical
}
\description{
Check if a variable is stored
}
