% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.empty}
\alias{is.empty}
\title{Testing whether a list is empty}
\usage{
is.empty(l)
}
\arguments{
\item{l}{A list}
}
\value{
TRUE or FALSE
}
\description{
Testing whether a list is empty
}
\keyword{internal}
