% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty_atom}
\alias{is_empty_atom}
\title{Testing for empty atomic vector.}
\usage{
is_empty_atom(l)
}
\arguments{
\item{l}{An object.}
}
\value{
TRUE or FALSE.
}
\description{
Testing for empty atomic vector.
}
\keyword{internal}
