% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{list_stored}
\alias{list_stored}
\title{List all stored variable names}
\usage{
list_stored()
}
\value{
Character vector of variable names
}
\description{
Returns the names of all variables currently stored in the parser environment.
}
