% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_parsers.R
\name{literal}
\alias{literal}
\title{The parser that matches an element using a literal string.}
\usage{
literal(string)
}
\arguments{
\item{string}{a single element character vector.}
}
\description{
\code{literal} tests whether a supplied string literally equals a desired value.
}
\section{Definition}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{literal(a)(x) <- satisfy(function(y) \{identical(y,a)\}) (x)
}\if{html}{\out{</div>}}

where \verb{= x} is to be understood as a function which tests its argument for
equality with \code{x}
}

\examples{
literal("ab") (c("ab", "cdef")) # success
literal("ab") (c("abc", "cdef")) # failure
}
