% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{print.marker}
\alias{print.marker}
\title{Prints an object of class \code{marker}}
\usage{
\method{print}{marker}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
An object of class \code{marker} is an empty list create by the function
\code{fail()}.
}
\details{
The \code{marker} class is used only for internal purposes. It contains the
attribute \code{n} which is the index of the element in which a parser failed.
Its \code{print} method is exported because the user of the \code{parcr} package will
often observe this object.
}
