% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_parsers.R
\name{split_by}
\alias{split_by}
\title{Applying a parser to a split string}
\usage{
split_by(p, split, finish = TRUE, fixed = FALSE, perl = FALSE)
}
\arguments{
\item{p}{A parser.}

\item{split}{A single element character vector  (or object which can be
coerced to such) containing \href{base::regex}{regular expression}(s)
(unless fixed = TRUE) to use for splitting. If empty matches occur, in
particular if split has length 0, x is split into single characters.}

\item{finish}{logical. Should the parser should completely consume the
string? Defaults to \code{TRUE}.}

\item{fixed}{
    logical.  If \code{TRUE} match \code{split} exactly, otherwise
    use regular expressions.  Has priority over \code{perl}.
  }

\item{perl}{logical.  Should Perl-compatible regexps be used?}
}
\value{
A parser.
}
\description{
Applying a parser to a split string
}
\details{
The function \code{\link[base:strsplit]{base::strsplit()}} is used to perform the splitting. The
parameters \code{split}, \code{fixed} and \code{perl} are inherited from that function.
}
