% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\docType{data}
\name{the}
\alias{the}
\alias{LNR}
\alias{set_LNR}
\alias{inc_LNR}
\alias{dec_LNR}
\alias{reset_LNR}
\title{The \code{the} environment is used for keeping track of the state variable \code{LNR}.}
\format{
An object of class \code{environment} of length 1.
}
\usage{
the

LNR()

set_LNR(n)

inc_LNR()

dec_LNR()

reset_LNR()
}
\arguments{
\item{n}{an \code{integer}, will be coerced to \code{integer}}
}
\description{
The \code{the} environment is used for keeping track of the state variable \code{LNR}.
}
\section{Functions}{
\itemize{
\item \code{LNR()}: Return the value of \code{LNR}.

\item \code{set_LNR()}: Set the value of \code{LNR} to \code{n}.

\item \code{inc_LNR()}: Increase the value of \code{LNR} by 1.

\item \code{dec_LNR()}: Decrease the value of \code{LNR} by 1.

\item \code{reset_LNR()}: Reset the value of \code{LNR} to 1.

}}
\keyword{internal}
