% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{Empty.line}
\alias{Empty.line}
\alias{Spacer}
\title{Recognize an empty line}
\usage{
Empty.line(cv)

Spacer(cv)
}
\arguments{
\item{cv}{A character vector}
}
\description{
An empty line is a line that consists entirely of space-like characters.
\code{Empty.line} is a parser that recognizes one empty line and \code{Spacer}
recognizes one or more empty lines.
}
\examples{
Empty.line (c(' \t  ')) # success
Empty.line (c('    .')) # failure
Spacer (c("   \t  ", "    ", "abc"))

}
