% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_parsers.R
\name{Empty.line}
\alias{Empty.line}
\alias{Spacer}
\alias{MaybeEmpty}
\title{Recognize empty lines}
\usage{
Empty.line()

Spacer()

MaybeEmpty()
}
\description{
An empty line is a line that consists entirely of space-like characters.
\code{Empty.line} is a parser that recognizes one empty line and \code{Spacer}
recognizes one or more empty lines and \code{MaybeEmpty} recognizes zero or more
empty lines. \code{Empty.line} actually returns the empty line but \code{Spacer} and
\code{MaybeEmpty} discard the empty lines.
}
\section{Definitions}{


\verb{Empty.line <- function() \{}

\code{satisfy(function(x) {stringr::str_replace_all(x, "\\\\s+", "") == ""})}

\verb{\}}

\code{Spacer <- function() {one.or.more(Empty.line()) \%ret\% NULL}}

\code{MaybeEmpty <- function() {(zero.or.more(Empty.line())) \%ret\% NULL}}
}

\examples{
Empty.line() (c(' \t  ')) # success
Empty.line() (c('    .')) # failure
Spacer() (c("   \t  ", "    ", "abc"))
Spacer() (c("            ", "    ", "Important text"))
Spacer() (c("Important text")) # failure
MaybeEmpty() (c("            ", "    ", "Important text"))
MaybeEmpty() (c("Important text")) # success, in contrast to Spacer()
}
