% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fastafile}
\alias{fastafile}
\title{Example nucleotide fasta file}
\format{
\subsection{\code{fastafile}}{

A character vector
}
}
\source{
Modified from \url{https://bioinformatics.org/annhyb/examples/seq_fasta.html}
}
\usage{
fastafile
}
\description{
An example fasta-formatted file with multiple nucleotide sequences.
}
\keyword{datasets}
