% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{finished}
\alias{finished}
\title{Test whether the parser has completely consumed the input}
\usage{
finished(o)
}
\arguments{
\item{o}{Output from applying a parser to a character vector.}
}
\value{
A logical value.
}
\description{
A parser has completely consumed its input when the input has satisfied
\code{eof()}.
}
\examples{
finished((literal("A") \%then\% eof())("A")) # TRUE
finished((literal("A"))("A")) # FALSE
finished((literal("A") \%then\% eof())(c("A","C"))) # FALSE
}
