% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{\%using\%}
\alias{\%using\%}
\title{Manipulate results from a parser by applying a function.}
\usage{
p \%using\% f
}
\arguments{
\item{p}{A parser}

\item{f}{A function to be applied to the result of a succesful \code{p}}
}
\value{
A parser
}
\description{
The \verb{\%using\%} combinator allows us to manipulate results from a parser, for
example building a parse tree. The parser \code{(p \%using\% f)} has the same
behavior as the parser \code{p}, except that the function \code{f} is applied to each
of its result values.
}
\section{Formal description}{


\verb{(p \%using$\% f) inp = [(fv, out) | (v, out) <- p inp]}
}

\examples{
(literal('ab') \%using\% toupper) (c("ab","cdef")) # success
(literal('ab') \%using\% toupper) (c("bb","cdef")) # failure
}
