% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.empty}
\alias{is.empty}
\title{Testing for empty atomic vector.}
\usage{
is.empty(l)
}
\arguments{
\item{l}{An object.}
}
\value{
TRUE or FALSE.
}
\description{
Testing for empty atomic vector.
}
\keyword{internal}
