% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{literal}
\alias{literal}
\title{The parser that matches an element using a literal string}
\usage{
literal(element)
}
\arguments{
\item{element}{A single-element (character) vector to be matched with the
first element of the right hand side}
}
\description{
\code{literal} is a parser for single elements. It tests whether the first
element in the vector is equal to a given element.
}
\section{Formal description}{


\verb{literal x = satisfy (= x)}

where \verb{= x} is to be understood as a function which tests its argument for
equality with \code{x}
}

\examples{
literal("ab") (c("ab", "cdef")) # success
literal("ab") (c("abc", "cdef")) # failure
}
