% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_parsers.R
\name{literal}
\alias{literal}
\title{The parser that matches an element using a literal string.}
\usage{
literal(string)
}
\arguments{
\item{string}{a single element character vector.}
}
\description{
\code{literal} tests whether a supplied string literally equals a desired value.
}
\section{Pseudocode}{

\preformatted{
literal(a) (x): satisfy(function(y) y==a)(x)
}
}

\examples{
literal("ab") (c("ab", "cdef")) # success
literal("ab") (c("abc", "cdef")) # failure
}
