% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_parsers.R
\name{Numbers}
\alias{Numbers}
\title{Extracts all integer and floating point numbers from a line}
\usage{
Numbers(n)
}
\arguments{
\item{n}{An integer.}
}
\value{
A parser.
}
\description{
Ignores any other symbols. It tests whether exactly n numbers are found.
}
\examples{
Numbers(3) ('1  2  3') #success
Numbers(3) ('1101\t201.5\t33') # success
Numbers(2) ('1101\t201.5\t33') # failure
Numbers(4) ('1101\t201.5\t33') # failure

}
