% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers_error.R
\name{Parser}
\alias{Parser}
\title{Turn a parser into an error messaging parser.}
\usage{
Parser(p)
}
\arguments{
\item{p}{a parser.}
}
\value{
A parser result or an error message about the line where the parser
failed.
}
\description{
Turn a parser into an error messaging parser.
}
\examples{
# Yields an error message about parser failing on line 5
\dontrun{
Parser(match_n(3,literal("a") \%then\% literal("t")) \%then\% eof())(c(rep(c("a","t"),2),"t","t"))
}
Parser(match_n(2,literal("a") \%then\% literal("t")) \%then\% eof())(rep(c("a","t"),2)) # success
}
