% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers_error.R
\name{reporter}
\alias{reporter}
\title{Turn a parser into an error reporting parser.}
\usage{
reporter(p)
}
\arguments{
\item{p}{a parser.}
}
\value{
The \code{L}-part of a successful parser result or an error message about
the line where the parser failed.
}
\description{
Turn a parser into an error reporting parser.
}
\examples{
# Yields an error message about parser failing on line 5
try(
  reporter(match_n(3,literal("a") \%then\%
   literal("t")) \%then\%
    eof())(c(rep(c("a","t"),2),"t","t"))
)
# No error, but parser result
reporter(match_n(2,literal("a") \%then\%
  literal("t")) \%then\%
  eof())(rep(c("a","t"),2))
}
