% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{satisfy}
\alias{satisfy}
\title{The parser that matches an element using a predicate function.}
\usage{
satisfy(b)
}
\arguments{
\item{b}{A boolean function to determine if the element is accepted.}
}
\description{
\code{satisfy} turns a predicate function into a parser that recognizes single
elements.
}
\section{Definition}{


\verb{satisfy(b)(x): fail()(x)             when x == list()}

\verb{             : succeed(x[1]) (x[-1]) when b(x[1])}

\verb{             : fail()(x)             when !b(x[1])}
}

\examples{

# define a predicate function that tests whether the next element starts
# with an 'a'
starts_with_a <- function(x) grepl("^a",x)
# Use it in the satisfy parser
satisfy(starts_with_a)(c("abc","def")) # success
satisfy(starts_with_a)(c("bca","def")) # failure
}
