% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{succeed}
\alias{succeed}
\title{The parser that always succeeds}
\usage{
succeed(left)
}
\arguments{
\item{left}{The part to be added on the L-side of a parsed list}
}
\description{
\code{succeed} parser always succeeds, without actually consuming any
input string. Since the outcome of succeed does not depend on its input, its
result value must be pre-determined, so it is included as an extra parameter.
}
\section{Formal description}{


\verb{succeed v inp = [(v, inp)].}
}

\keyword{internal}
